Project 12 with challenges
Updated the FilteredList struct with several init parameters covering the predicate to use, case sensitivity, negation, sort order, and the choice of firstName versus lastName as the key to use in predicates and sorting. Added some controls and wrapped it all up in a Form inside a NavigationView to make it easy to exercise the various FilteredView options.